// src/models/Statistic.ts
export type Statistic = {
    active_carrier: number;
    active_client: number;
    aggregate_date: string;
    assigned_count: number;
    margin_abs: number;
    margin_abs_per_order: number;
    margin_perc: number;
    new_carriers: number;
    new_clients: number;
    order_count: number;
    order_per_period: number;
    revenue: number;
    revenue_assigned: number;
    revenue_per_order: number;
  };